Remove redundant casts of literal binary operators on the LHS.
`(==|\+|-|\*|/|>=|<=|<<|>>| <| >|&|\|) ([0-9]+) as libc::c_int` => `$1 $2`

This is a repeat of afc35e2 after #425.

+++ tests/seek_stress.rs
@@ -229,8 +229,7 @@ unsafe extern "C" fn decode_rand(
         r#ref: 0 as *mut Dav1dRef,
         allocator_data: 0 as *mut libc::c_void,
     };
-    let num_frames: libc::c_int =
-        xor128_rand() % (fps * 5 as libc::c_int as libc::c_double) as libc::c_int;
+    let num_frames: libc::c_int = xor128_rand() % (fps * 5 as libc::c_double) as libc::c_int;
     let mut i = 0;
     while i < num_frames {
         res = decode_frame(&mut p, c, data);
@@ -469,7 +468,7 @@ unsafe fn main_0(argc: libc::c_int, argv: *const *mut libc::c_char) -> libc::c_i
         i_fps.as_mut_ptr(),
         &mut total,
         i_timebase.as_mut_ptr(),
-    ) < 0 as libc::c_int
+    ) < 0
         || i_timebase[0] == 0
         || i_timebase[1] == 0
         || i_fps[0] == 0
@@ -486,7 +485,7 @@ unsafe fn main_0(argc: libc::c_int, argv: *const *mut libc::c_char) -> libc::c_i
     if !(fps < 1 as libc::c_double) {
         let mut i: libc::c_int = 0;
         loop {
-            if !(i < 3 as libc::c_int) {
+            if !(i < 3) {
                 current_block = 5948590327928692120;
                 break;
             }
@@ -511,13 +510,11 @@ unsafe fn main_0(argc: libc::c_int, argv: *const *mut libc::c_char) -> libc::c_i
                 let mut i_0 = 0;
                 let mut tries = 0;
                 loop {
-                    if !(i_0 - tries < 4 as libc::c_int
-                        && tries < 4 as libc::c_int / 2 as libc::c_int)
-                    {
+                    if !(i_0 - tries < 4 && tries < 4 / 2) {
                         current_block = 8693738493027456495;
                         break;
                     }
-                    let sign: libc::c_int = if xor128_rand() & 1 as libc::c_int != 0 {
+                    let sign: libc::c_int = if xor128_rand() & 1 != 0 {
                         -(1 as libc::c_int)
                     } else {
                         1 as libc::c_int
@@ -597,7 +594,7 @@ unsafe fn main_0(argc: libc::c_int, argv: *const *mut libc::c_char) -> libc::c_i
                         }
                         // simulate seeking after the end of the file
                         let mut i_1: libc::c_int = 0;
-                        while i_1 < 2 as libc::c_int {
+                        while i_1 < 2 {
                             if seek(
                                 in_0,
                                 c,
